refactor: server의 책임 분담
- RouterManager로 라우팅 로직 분리
- ServerConfiguration로 서버 세팅 로직 분리
- 복잡한 의존성 주입 제거 및 사용하지 않는 객체 제거
